styles updated

+++ server/api/controllers/note.controller.js
@@ -4,7 +4,11 @@ import { errorHandler } from "../utils/error.js";
 // Create a new note
 export const createNote = async (req, res, next) => {
   const { title, content } = req.body;
-  //console.log(req.user);
+
+  if (!title || !content) {
+    return next(errorHandler(400, 'Title and content are required'));
+  }
+
   try {
     const newNote = new Note({
       title,
@@ -32,7 +36,18 @@ export const getNotesByUserId = async (req, res, next) => {
 export const updateNote = async (req, res, next) => {
   const { id } = req.params;
   const { title, content } = req.body;
+
   try {
+    const note = await Note.findById(id);
+
+    if (!note) {
+      return next(errorHandler(404, 'Note not found'));
+    }
+
+    if (note.userId !== req.user.id) {
+      return next(errorHandler(403, 'You are not authorized to update this note'));
+    }
+
     const updatedNote = await Note.findByIdAndUpdate(
       id,
       { title, content },
@@ -44,10 +59,22 @@ export const updateNote = async (req, res, next) => {
   }
 };
 
+
 // Delete a note
 export const deleteNote = async (req, res, next) => {
   const { id } = req.params;
+
   try {
+    const note = await Note.findById(id);
+
+    if (!note) {
+      return next(errorHandler(404, 'Note not found'));
+    }
+
+    if (note.userId !== req.user.id) {
+      return next(errorHandler(403, 'You are not authorized to delete this note'));
+    }
+
     await Note.findByIdAndDelete(id);
     res.status(200).json({ message: "Note deleted successfully!" });
   } catch (error) {
